requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.2
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.9
django-cors-headers==3.10.1
djangorestframework==3.12.4
idna==3.3
kombu==5.2.4
prompt-toolkit==3.0.30
psycopg2-binary==2.9.1
pytz==2022.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
